// src/interfaces/global/shoppingCar.ts
export interface ShoppingCarType {
    id: number
    CreatedAt: string
    UpdatedAt: string
    DeletedAt: any
    name: string
    description: string
    price: number
    photo: string
    available: boolean
    company_id: number
    category_product_id: number
    comments: any
    amount: number
  }
  